Revert change to opttools.
